feat: productImage 추가/수정 및 상품 상세페이지에서 조회 가능 (#343)
* feat: default 이미지 추가

* feat: 하트 위치, 화면 폭에 맞게 유연하게 조정

* feat: imageUrl 추가/삭제 기능

* feat: imageUrl 추가/삭제 기능

* feat: useProductImage 개선

* feat: imageUrl 추가/삭제 기능

* fix: catch 구문 작성

* refactor: addImage 시 alert 제거

* feat: imageUrl 추가/삭제 기능

* fix: conflict 해결

* fix: fetch 문제, controlled 문제 해결

* fix: defaultValue 수정, initialProductFormState 문제 해결

* fix: wishList 커스텀 훅을 조건부로 호출하는 오류 수조ㅓㅇ

* fix: wishList 커스텀 훅을 조건부로 호출하는 오류 수정

* feat: 상품 상세페이지에서 모든 이미지 볼 수 있도록

* refator

* refator

+++ frontend/src/context/ProductsContext.tsx
@@ -62,10 +62,19 @@ export const useProductItem = (id: number) => {
   const [product, setProduct] = useState<Product | null>(null);
 
   useEffect(() => {
-    apiFetch(`/api/v1/products/${id}`).then(setProduct);
+    if(!id || id === 0) {
+      return;
+    }
+    apiFetch(`/api/v1/products/${id}`).then(setProduct)
+    .catch((error) => {
+      alert(`${error.resultCode} : ${error.msg}`);
+    });
   }, []);
 
   const deleteProduct = (onSuccess: (data: any) => void) => {
+    if(!id || id === 0) {
+      return;
+    }
     apiFetch(`/api/v1/adm/products/${id}`, {
       method: "DELETE"
     }).then(onSuccess)
@@ -82,6 +91,9 @@ export const useProductItem = (id: number) => {
     imageUrl: string,
     onSuccess:(data: any) => void
   }) => {
+    if(!id || id === 0) {
+      return;
+    }
     apiFetch(`/api/v1/adm/products/${id}`, {
       method: "PUT",
       body: JSON.stringify({
@@ -100,8 +112,73 @@ export const useProductImage = (productId: number) => {
   const [productImages, setProductImages] = useState<ProductImage[] | null>(null);
 
   useEffect(() => {
-    apiFetch(`/api/v1/products/${productId}/images`).then(setProductImages);
+    if(!productId || productId === 0) {
+      return;
+    }
+    apiFetch(`/api/v1/products/${productId}/images`).then(setProductImages)
+    .catch((error) => {
+      alert(`${error.resultCode} : ${error.msg}`);
+    });
+  }, [productId]);
+
+  const addProductImage = (url: string, productId: number) => {
+    if(!productId || productId === 0) {
+      return;
+    }
+    apiFetch(`/api/v1/adm/products/${productId}/images`, {
+      method: "POST",
+      body: JSON.stringify({url}),
+    }).then()
+    .catch((error) => {
+      alert(`${error.resultCode} : ${error.msg}`);
+    });
+  }
+
+  const firstImage = productImages?.[0]?.url;
+
+  return { productImages, addProductImage, firstImage };
+}
+
+export const useProductImageItem = (productId: number, id: number) => {
+  const [productImage, setProductImage] = useState<ProductImage>();
+
+  useEffect(() => {
+    if(!productId || productId === 0) {
+      return;
+    }
+    apiFetch(`/api/v1/products/${productId}/images/${id}`).then(setProductImage)
+    .catch((error) => {
+      alert(`${error.resultCode} : ${error.msg}`);
+  });
   }, []);
 
-  return productImages?.[0]?.url;
+  const deleteProductImage = (onSuccess: (data: any) => void) => {
+    if(!productId || productId === 0) {
+      return;
+    }
+    apiFetch(`/api/v1/adm/products/${productId}/images/${id}`, {
+      method: "DELETE"
+    }).then(onSuccess)
+    .catch((error) => {
+      alert(`${error.resultCode} : ${error.msg}`);
+    });
+  };
+
+  const modifyProductImage = ({url, onSuccess}: {
+    url: string,
+    onSuccess:(data: any) => void
+  }) => {
+    if(!productId || productId === 0) {
+      return;
+    }
+    apiFetch(`/api/v1/adm/products/${productId}/images/${id}`, {
+      method: "PUT",
+      body: JSON.stringify({url}),
+    }).then(onSuccess)
+    .catch((error) => {
+      alert(`${error.resultCode} : ${error.msg}`);
+    });
+  }
+
+  return {productImage, deleteProductImage, modifyProductImage};
 }
